// src/products/enums/product-category.enum.ts
export enum ProductCategory {
  ELECTRONICS = 'Electronics',
  FASHION = 'Fashion',
  HOME = 'Home',
  BOOKS = 'Books',
  TOYS = 'Toys',
  SPORTS = 'Sports',
  OUTDOORS = 'Outdoors',
  HEALTH = 'Health',
  BEAUTY = 'Beauty',
  GROCERY = 'Grocery',
  AUTOMOTIVE = 'Automotive',
  OFFICE_SUPPLIES = 'Office Supplies',
  PETS = 'Pets',
  BABY = 'Baby',
  MUSIC = 'Music',
  GARDEN = 'Garden',
  DIY = 'DIY',
  SOFTWARE = 'Software',
  GAMING = 'Gaming',
  MOVIES = 'Movies',
  JEWELRY = 'Jewelry',
  FURNITURE = 'Furniture',
  KITCHEN = 'Kitchen',
  CLEANING = 'Cleaning',
  CLOTHING = 'Clothing',
  SHOES = 'Shoes',
  BAGS = 'Bags',
  WATCHES = 'Watches',
  ART = 'Art',
  CRAFTS = 'Crafts',
  TRAVEL = 'Travel',
  PHOTOGRAPHY = 'Photography',
  FITNESS = 'Fitness',
  SECURITY = 'Security',
  APPLIANCES = 'Appliances',
  MOBILE_PHONES = 'Mobile Phones',
  COMPUTERS = 'Computers',
  CAMERAS = 'Cameras',
  AUDIO = 'Audio',
  OFFICE_FURNITURE = 'Office Furniture',
  LIGHTING = 'Lighting',
  HOLIDAY_DECOR = 'Holiday Decor',
  COLLECTIBLES = 'Collectibles',
  LUGGAGE = 'Luggage',
  MEDICAL_SUPPLIES = 'Medical Supplies',
  EDUCATIONAL = 'Educational',
  PARTY_SUPPLIES = 'Party Supplies',
  DRONES = 'Drones',
  HOBBIES = 'Hobbies',
  ANTIQUES = 'Antiques',
  WINE = 'Wine',
  SPIRITS = 'Spirits',
  CIGARS = 'Cigars',
  VAPES = 'Vapes',
  BUILDING_SUPPLIES = 'Building Supplies',
  SAFETY = 'Safety',
  TELEVISIONS = 'Televisions',
  SMALL_APPLIANCES = 'Small Appliances',
  PERSONAL_CARE = 'Personal Care',
  INSTRUMENTS = 'Instruments',
  MOTORCYCLES = 'Motorcycles',
  INDUSTRIAL = 'Industrial',
  WATER_SPORTS = 'Water Sports',
  WINTER_SPORTS = 'Winter Sports',
  CYCLING = 'Cycling',
  RUNNING = 'Running',
  VIDEO_GAMES = 'Video Games',
  BOARD_GAMES = 'Board Games',
  CONSTRUCTION = 'Construction',
  COMMERCIAL_SUPPLIES = 'Commercial Supplies',
  ENERGY = 'Energy',
  FISHING = 'Fishing',
  CAMPING = 'Camping',
  BOATING = 'Boating',
  AIR_SOFT = 'Airsoft',
  MODEL_KITS = 'Model Kits',
  SCUBA = 'Scuba',
  AUDIOBOOKS = 'Audiobooks',
  MAGAZINES = 'Magazines',
  CHILDRENS_TOYS = 'Children’s Toys',
  PUZZLES = 'Puzzles',
  COSMETICS = 'Cosmetics',
  PERFUMES = 'Perfumes',
  BEVERAGES = 'Beverages',
}